Adjust some regex debugging printouts to not give wrong-format-width warnings on a 64-bit machine.  Noted while chasing a recent regex bug report.

+++ src/backend/regex/regcomp.c
@@ -2132,7 +2132,7 @@ dump(regex_t *re,
 
 	fprintf(f, "\n\n\n========= DUMP ==========\n");
 	fprintf(f, "nsub %d, info 0%lo, csize %d, ntree %d\n",
-			re->re_nsub, re->re_info, re->re_csize, g->ntree);
+			(int) re->re_nsub, re->re_info, re->re_csize, g->ntree);
 
 	dumpcolors(&g->cmap, f);
 	if (!NULLCNFA(g->search))
